requirements.txt
# python -> 3.11.13
numpy
pandas
bs4
requests
fastapi
fastapi[standard]
python-dotenv
